+++ src/GWind/BillingProject/Domain.hs
@@ -1,9 +1,11 @@
 -- ⚠
-module GWind.BillingProject.Domain (BillingSnake)
+module GWind.BillingProject.Domain (BillingSnake, BillingAccountName, BillingAccountDisplayName)
 where
 
     import GWind.Common.Domain (Snake)
 
     type BillingSnake = Snake
+    type BillingAccountName = String
+    type BillingAccountDisplayName = String
 
 --